+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Refactor fields renderer to support custom actions on fields
 - Update Vue to version 3.3.7
 - Sort content models alphabetically on overview page
+- Add option to store content models in database
 
 ## 2.7.1 (2023-10-08)